requirements.txt
python-dotenv
google-api-python-client
youtube-transcript-api
nltk
yt-dlp
matplotlib
seaborn
pandas
wordcloud